+++ metron-interface/metron-rest/src/main/java/org/apache/metron/rest/controller/PcapSampleController.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.metron.rest.service.impl;
+package org.apache.metron.rest.controller;
 
 import com.google.common.collect.Lists;
 import org.apache.hadoop.conf.Configuration;
@@ -23,22 +23,27 @@ import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.metron.common.hadoop.SequenceFileIterable;
 import org.apache.metron.common.utils.timestamp.TimestampConverters;
-import org.apache.metron.pcap.PcapMerger;
 import org.apache.metron.pcap.filter.query.QueryPcapFilter;
 import org.apache.metron.pcap.mr.PcapJob;
-import org.apache.metron.rest.model.PcapsResponse;
-import org.apache.metron.rest.service.PcapQueryService;
-import org.springframework.stereotype.Service;
+import org.apache.metron.rest.RestException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
-@Service
-public class PcapQueryServiceImpl implements PcapQueryService {
+// NOTE:  This is meant only as an example and should be replaced as part of https://issues.apache.org/jira/browse/METRON-1559
+@RestController
+@RequestMapping("/api/v1/pcap/sample")
+public class PcapSampleController {
 
-  @Override
-  public PcapsResponse query() {
+  @RequestMapping(method = RequestMethod.GET)
+  ResponseEntity<PcapsResponse> query() throws RestException {
     PcapsResponse response = new PcapsResponse();
     PcapJob pcapJob = new PcapJob();
     Configuration configuration = new Configuration();
@@ -58,16 +63,66 @@ public class PcapQueryServiceImpl implements PcapQueryService {
     } catch (IOException | ClassNotFoundException | InterruptedException e) {
       e.printStackTrace();
     }
-    return response;
+    return new ResponseEntity<>(response, HttpStatus.OK);
   }
 
-  protected byte[] mergePcaps(List<byte[]> pcaps) throws IOException {
-    if (pcaps.size() == 1) {
-      return pcaps.get(0);
-    }
-    ByteArrayOutputStream baos = new ByteArrayOutputStream();
-    PcapMerger.merge(baos, pcaps);
-    return baos.toByteArray();
+  /**
+   * Holds pcaps data, status and the partial response key.
+   *
+   * @author Sayi
+   */
+  class PcapsResponse {
 
+    /** response of the processed keys. */
+    private List<byte[]> pcaps = new ArrayList<byte[]>();;
+
+    /**
+     * Sets the pcaps.
+     *
+     * @param pcaps
+     *          the new pcaps
+     */
+    public void setPcaps(List<byte[]> pcaps) {
+      this.pcaps = pcaps;
+    }
+
+    /**
+     * Adds the pcaps.
+     *
+     * @param pcaps
+     *          the pcaps
+     */
+    public void addPcaps(byte[] pcaps) {
+      this.pcaps.add(pcaps);
+    }
+
+
+    /**
+     * Gets the response size.
+     *
+     * @return the response size
+     */
+    public long getResponseSize() {
+      long responseSize = 0;
+      for (byte[] pcap : this.pcaps) {
+        responseSize = responseSize + pcap.length;
+      }
+      return responseSize;
+    }
+
+    /**
+     * Gets the pcaps.
+     *
+     * @return the pcaps
+     * @throws IOException
+     *           Signals that an I/O exception has occurred.
+     */
+    public List<byte[]> getPcaps() throws IOException {
+      if(pcaps == null) {
+        return new ArrayList<>();
+      } else {
+        return pcaps;
+      }
+    }
   }
 }
